BugzID:220 Change ints to longs.

+++ com-mmxlabs-lng-scheduler/com.mmxlabs.scheduler.optimiser/src/test/java/com/mmxlabs/scheduler/optimiser/fitness/components/CostComponentTest.java
@@ -65,15 +65,15 @@ public class CostComponentTest {
 
 		// define the fuel consumptions
 		// multiply by scale factor (i.e. convert to internal unit)
-		final int baseConsumption = 1000 * Calculator.ScaleFactor;
-		final int FBOConsumption = 0 * Calculator.ScaleFactor;
-		final int NBOConsumption = 500 * Calculator.ScaleFactor;
+		final long baseConsumption = 1000 * Calculator.ScaleFactor;
+		final long FBOConsumption = 0 * Calculator.ScaleFactor;
+		final long NBOConsumption = 500 * Calculator.ScaleFactor;
 
 		// define the fuel unit prices
 		// multiply by scale factor (i.e. convert to internal unit)
-		final int baseUnit = 13 * Calculator.ScaleFactor;
-		final int NBOUnit = 25 * Calculator.ScaleFactor;
-		final int FBOUnit = 7 * Calculator.ScaleFactor;
+		final long baseUnit = 13 * Calculator.ScaleFactor;
+		final long NBOUnit = 25 * Calculator.ScaleFactor;
+		final long FBOUnit = 7 * Calculator.ScaleFactor;
 
 		final String name = "name";
 		final CargoSchedulerFitnessCore<Object> core = null;
@@ -91,9 +91,9 @@ public class CostComponentTest {
 		voyage.setFuelConsumption(FuelComponent.NBO, FuelComponent.NBO.getDefaultFuelUnit(), NBOConsumption);
 		voyage.setFuelConsumption(FuelComponent.FBO, FuelComponent.FBO.getDefaultFuelUnit(), FBOConsumption);
 		// set unit prices
-		voyage.setFuelUnitPrice(FuelComponent.Base, baseUnit);
-		voyage.setFuelUnitPrice(FuelComponent.NBO, NBOUnit);
-		voyage.setFuelUnitPrice(FuelComponent.FBO, FBOUnit);
+		voyage.setFuelUnitPrice(FuelComponent.Base, (int)baseUnit);
+		voyage.setFuelUnitPrice(FuelComponent.NBO, (int)NBOUnit);
+		voyage.setFuelUnitPrice(FuelComponent.FBO, (int)FBOUnit);
 
 		final Object[] voyageSequence = new Object[] { voyage };
 		final VoyagePlan voyagePlan = new VoyagePlan();
@@ -110,11 +110,11 @@ public class CostComponentTest {
 		final long cost = c.endEvaluationAndGetCost();
 
 		// divide by scale factor to convert from internal unit to external (1000$ to $)
-		final int expectedBaseCost = (baseConsumption / Calculator.ScaleFactor) * (baseUnit / Calculator.ScaleFactor);
-		final int expectedNBOCost = (NBOConsumption / Calculator.ScaleFactor) * (NBOUnit / Calculator.ScaleFactor);
-		final int expectedFBOCost = (FBOConsumption / Calculator.ScaleFactor) * (FBOUnit / Calculator.ScaleFactor);
+		final long expectedBaseCost = (baseConsumption / Calculator.ScaleFactor) * (baseUnit / Calculator.ScaleFactor);
+		final long expectedNBOCost = (NBOConsumption / Calculator.ScaleFactor) * (NBOUnit / Calculator.ScaleFactor);
+		final long expectedFBOCost = (FBOConsumption / Calculator.ScaleFactor) * (FBOUnit / Calculator.ScaleFactor);
 
-		final int expectedTotalCost = expectedBaseCost + expectedFBOCost + expectedNBOCost;
+		final long expectedTotalCost = expectedBaseCost + expectedFBOCost + expectedNBOCost;
 
 		Assert.assertEquals("Expected cost equals calculated cost.", expectedTotalCost, cost);
